% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdc}
\alias{cdc}
\title{CDC data on suicide attempt and mortality rates}
\format{\preformatted{'data.frame':	36 obs. of  6 variables:
 $ sex         : Factor w/ 2 levels "Females","Males": 2 2 2 2 2 2 2 2 2 2 ...
 $ age         : Factor w/ 18 levels "00-04 yrs","05-09 yrs",..: 1 2 3 4 5 6 7 8 9 10 ...
 $ age2        : num  2.5 7.5 12.5 17.5 22.5 27.5 32.5 37.5 42.5 47.5 ...
 $ population  : int  111871065 111734405 118222191 121559927 117990967 110795467 110147304 114664994 120430411 120486134 ...
 $ Suicide rate: num  0 0.04 1.65 11.8 21.08 ...
 $ Attempt rate: num  0.163 3.461 40.576 206.454 248.64 ...
}}
\usage{
cdc
}
\description{
United States 2001-2011
}
\keyword{datasets}

