% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coding}
\alias{coding}
\title{Coding of variables on all suicide reports from the probability sample.}
\format{\preformatted{'data.frame':	474 obs. of  30 variables:
 $ id                             : chr  "1" "2" "3" "4" ...
 $ culture                        : chr  "Akan" "Akan" "Akan" "Akan" ...
 $ record                         : chr  "At first the Juabens were so successful that [p. 55] many of the Kumasi captains, among whom were the Adumhene, the Anantahene,"| __truncated__ "A defeated Ashanti captain was always expected to commit suicide. Death, to the Ashanti, was preferable to disgrace." "As regards the suicide, the presumption was that he had killed himself because he had committed a sin which was a taboo and, in"| __truncated__ "it was that of telling or expecting the person found guilty to commit suicide.(42) The same Rev. Mader reports another case of "| __truncated__ ...
 $ record_type                    : chr  "Case" "Cultural model" "Case" "Case" ...
 $ age_category                   : Factor w/ 4 levels "Child","Adolescent",..: 3 3 3 3 3 3 3 3 3 3 ...
 $ sex                            : Factor w/ 4 levels "both","female",..: 3 3 3 2 4 2 4 3 3 4 ...
 $ sb_type                        : Factor w/ 5 levels "completion","attempt",..: 1 1 1 1 1 2 1 1 1 1 ...
 $ Conflict                       : int  1 1 0 0 0 1 0 0 1 0 ...
 $ Fitness threat                 : int  1 1 0 0 1 1 0 0 1 0 ...
 $ High RP                        : int  1 1 0 0 0 0 0 1 1 0 ...
 $ Low lethality of attempt       : int  0 0 0 0 0 1 0 0 0 0 ...
 $ Motive: Leverage               : int  0 0 0 0 0 0 0 0 0 0 ...
 $ Outcome: victim better off     : int  0 0 0 0 0 0 0 0 0 0 ...
 $ Social partner fitness         : int  0 0 0 0 0 0 0 0 0 0 ...
 $ Powerlessness                  : int  1 1 0 1 1 0 0 0 1 0 ...
 $ Motive: harm others            : int  0 0 0 0 0 0 0 0 0 0 ...
 $ Outcome: others harmed         : int  0 0 0 1 0 0 0 0 0 0 ...
 $ AngrySB                        : int  1 1 0 0 0 0 0 0 0 0 ...
 $ Public SB                      : int  1 0 0 0 0 1 0 0 0 0 ...
 $ Low RP                         : int  0 0 0 0 0 1 0 0 0 0 ...
 $ Motive: kin better off         : int  0 0 0 0 0 0 0 0 0 0 ...
 $ Motive: make others better off : int  0 0 0 0 0 0 0 0 0 0 ...
 $ Evidence of burdensomeness     : int  0 0 0 0 0 1 0 0 0 0 ...
 $ Outcome: others better off     : int  0 0 0 0 0 0 0 0 0 0 ...
 $ Outcome: victim worse off      : int  0 0 0 0 0 0 0 0 0 0 ...
 $ High lethality of attempt      : int  1 0 0 1 0 0 0 0 0 0 ...
 $ No social partner fitness      : int  1 1 0 0 0 0 0 0 1 0 ...
 $ Private SB                     : int  0 0 0 1 0 0 0 0 0 0 ...
 $ Evidence against burdensomeness: int  0 0 0 0 0 0 0 0 0 0 ...
 $ Evidence against fitness threat: int  0 0 0 0 0 0 0 0 0 0 ...
}}
\usage{
coding
}
\description{
Coding of 23 variables related to the inclusive fitness and bargaining models of suicidal behavior.
}
\details{
Kristen Syme, Zach Garfield and Ed Hagen
}
\keyword{datasets}

