% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cultures}
\alias{cultures}
\title{Cultures from the HRAF Probability Sample with information on suicide}
\format{\preformatted{'data.frame':	57 obs. of  10 variables:
 $ Culture.code         : chr  "AA01" "AD05" "AO07" "AR05" ...
 $ Name                 : chr  "Korea" "Taiwan Hokkien" "Central Thai" "Garo" ...
 $ Region               : chr  "East Asia" "East Asia" "Southeast Asia" "South Asia" ...
 $ Continental_region   : chr  "Asia" "Asia" "Asia" "Asia" ...
 $ subsistence          : Factor w/ 5 levels "hunter-gatherers",..: 4 4 4 2 4 4 1 4 3 5 ...
 $ complexity           : num  48 27 50 31 31 NA 14 36 18 47 ...
 $ postmarital.residence: Factor w/ 5 levels "Neolocal","Ambilocal",..: 3 3 5 5 3 NA 2 3 3 3 ...
 $ descent              : Factor w/ 5 levels "Bilateral","Ambilineal",..: 3 2 1 5 3 NA 1 3 1 1 ...
 $ longitude            : num  128 121 102 90 85.5 80 92.5 21.5 25 -5 ...
 $ latitude             : num  38 24 13.5 25.5 23 8 12 43.5 64.5 56 ...
}}
\usage{
cultures
}
\description{
Contains information on cultures in the HRAF probability sample, including
information from matching SCCS cultures.
}
\keyword{datasets}

